Cuti

+++ src/main/java/bta/cabang/operasional/controller/CutiController.java
@@ -12,17 +12,18 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 
 @Controller
 public class CutiController {
     @Autowired
     private CutiService cutiService;
-
-    @Autowired
-    private UserService userService;
 
     @Autowired
     private AuthService authService;
@@ -35,7 +36,7 @@ public class CutiController {
 
         List<CutiModel> listCuti = new ArrayList<>();
         if (role == 1 || role == 2) {
-            listCuti = cutiService.getAllCuti();
+            listCuti = cutiService.getAllCutiByStatus(0);
         } else {
             listCuti = cutiService.getAllCutiByUser(id);
         }
@@ -50,22 +51,21 @@ public class CutiController {
     @GetMapping("/cuti/add")
     public String addCutiForm(Model model) {
         model.addAttribute("cuti", new CutiModel());
-
         return "form-addCuti";
     }
 
     @PostMapping("/cuti/add")
     public String addCutiSubmit(@ModelAttribute CutiModel cuti, RedirectAttributes redirectAttrs){
-//        try {
+        try {
             cuti.setPengaju(authService.getCurrentLoggedInUserByUsername());
             cutiService.addCuti(cuti);
             redirectAttrs.addFlashAttribute("alert", "addSuccess");
             return "redirect:/cuti";
 
-//        } catch (Exception e) {
-//            redirectAttrs.addFlashAttribute("alert", "addFail");
-//            return "redirect:/cuti";
-//        }
+        } catch (Exception e) {
+            redirectAttrs.addFlashAttribute("alert", "addFail");
+            return "redirect:/cuti";
+        }
 
     }
 
@@ -77,22 +77,34 @@ public class CutiController {
             Long role = currentUser.getRole().getIdRole();
 
             model.addAttribute("cuti", cuti);
-            model.addAttribute("isAbleToAddDeleteCuti", role == 3 || role == 4 || role ==5);
+            model.addAttribute("isPegawai", role == 3 || role == 4 || role ==5);
             model.addAttribute("isAbleToUpdateCuti", role == 1 || role == 2);
             return "view-cuti";
 
-        } catch (EmptyResultDataAccessException e) {
+        } catch (NoSuchElementException e) {
             redirectAttrs.addFlashAttribute("alert", "notFound");
             return "redirect:/cuti";
         }
     }
 
+    @PostMapping("/cuti/update")
+    public String updateStatusCuti(@ModelAttribute CutiModel cuti, RedirectAttributes redirectAttrs) {
+        cutiService.updateCuti(cuti);
+        redirectAttrs.addFlashAttribute("alert", "updateSuccess");
+        return "redirect:/cuti";
+    }
+
     @GetMapping("cuti/delete/{id}")
     private String deleteCuti(@PathVariable Long id, RedirectAttributes redirectAttrs) {
         try {
-            cutiService.deleteCuti(id);
-            redirectAttrs.addFlashAttribute("alert", "delSuccess");
-            return "redirect:/cuti";
+            if (cutiService.getCutiByIdCuti(id).getStatus() == 0) {
+                cutiService.deleteCuti(id);
+                redirectAttrs.addFlashAttribute("alert", "delSuccess");
+                return "redirect:/cuti";
+            } else {
+                redirectAttrs.addFlashAttribute("alert", "delNotAllowed");
+                return "redirect:/cuti";
+            }
 
         } catch (EmptyResultDataAccessException e) {
             redirectAttrs.addFlashAttribute("alert", "notFound");
